WIP: Creating interfaces, splitting code into modules
